backend/requirements.txt
# ============================================================
# Talk2SQL Backend — Python Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# Python:  3.11.2+
# ============================================================

# ── Web Framework ───────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database (Supabase / PostgreSQL) ────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6

# ── LLM Client (OpenAI-compatible for LongCat) ─────────────
openai==1.58.1

# ── Auth ────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── CORS / Utils ────────────────────────────────────────────
python-dotenv==1.0.1

# ── Logging / Dev ───────────────────────────────────────────
rich==13.9.4
httpx==0.28.1
